필터링 데이터 뷰엑스 설계 완료 및 날짜별 필터부분 수정중
